pyproject.toml
[tool.poetry]
name = "4"
version = "0.1.0"
description = ""
authors = ["Etienne SENIGOUT <etienne.senigout@epita.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pygame = {version = "^2.5.2", allow-prereleases = true}
numpy = ">=2.1.0"
matplotlib = {version = "^3.8.0", allow-prereleases = true}
gymnasium = {extras = ["other"], version = "^1.2.1"}
torch = "^2.9.0"
ale-py = "^0.11.2"
torchvision = "^0.24.0"
array-api-compat = "^1.12.0"
pyaml = "^25.7.0"
torchrl = "^0.10.1"
rich = "^14.2.0"
pandas = "^2.3.3"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
